+++ packages/ruleset-bundler/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@stoplight/spectral-ruleset-bundler-v1.3.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ruleset-bundler-v1.3.0...@stoplight/spectral-ruleset-bundler-v1.3.1) (2022-08-03)
+
+
+### Bug Fixes
+
+* **ruleset-bundler:** never externalize builtins ([#2174](https://github.com/stoplightio/spectral/issues/2174)) ([fb1bbe6](https://github.com/stoplightio/spectral/commit/fb1bbe61278d40c7aefb946997663b360863bae2))
+
 # [@stoplight/spectral-ruleset-bundler-v1.3.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-ruleset-bundler-v1.2.1...@stoplight/spectral-ruleset-bundler-v1.3.0) (2022-06-30)